Cache basic tests

+++ packages/core/src/cache.ts
@@ -4,31 +4,40 @@ import fs = require('fs')
 import { Config } from './config'
 import { Injectable } from './decorators/injectable'
 
+export interface CacheOpts {
+    dir: string
+    disabled: boolean
+}
+
 @Injectable()
 export class Cache {
 
-    dir: any
-    opts: any
+    dir: string
+    opts: CacheOpts
 
     constructor(config: Config) {
         this.opts = config.get('cache', {})
-        this.dir = path.join(process.cwd(), 'storage/cache')
+        if (!this.opts.dir) {
+            throw new Error('Cache target folder must be set')
+        }
+        this.dir = path.resolve(this.opts.dir)
     }
 
-    async get(key: string, cb: () => any, opts?: any): Promise<any> {
+    async get(key: string, cb?: () => any, opts?: any): Promise<any> {
         const hashed = this.hash(key)
         if (this.opts.disabled) {
             return cb()
         }
         try {
-            const contents = fs.readFileSync(path.join(this.dir, hashed), 'utf8')
+            const contents = await fs.promises.readFile(path.join(this.dir, hashed), 'utf8')
             const now = Math.floor(Date.now() / 1000)
             const time = Number(contents.slice(0, 10))
+
             if (time < now) {
                 this.flush(key)
                 if (cb) {
                     const data = await cb()
-                    this.put(key, data, opts.ttl)
+                    await this.put(key, data, opts.ttl)
                     return data
                 }
                 return null
@@ -38,7 +47,7 @@ export class Cache {
         } catch (e) {
             if (cb && typeof cb === 'function') {
                 const data = await cb()
-                this.put(key, data)
+                await this.put(key, data)
                 return data
             }
             return null
@@ -53,17 +62,17 @@ export class Cache {
     async flushAll() {
         const promises: Promise<any>[] = await new Promise((resolve, reject) => {
             fs.readdir(this.dir, (err, contents) => {
-                if(err) return reject(err)
+                if (err) return reject(err)
                 resolve(contents.map(file => this.flushFile(file)))
             })
         })
         return Promise.all(promises)
     }
 
-    put(key: string, value: any, time?: number) {
+    async put(key: string, value: any, time?: number) {
         const hashed = this.hash(key)
         const val = JSON.stringify(value)
-        fs.writeFileSync(path.join(this.dir, hashed), this.getExpiration(time) + val, 'utf8')
+        return fs.promises.writeFile(path.join(this.dir, hashed), this.getExpiration(time) + val, 'utf8')
     }
 
     hash(key: string) {
@@ -79,8 +88,11 @@ export class Cache {
 
     private flushFile(file): Promise<void> {
         return new Promise((resolve, reject) => {
+            if (path.extname(file) !== '') {
+                return resolve()
+            }
             fs.unlink(path.join(this.dir, file), (err) => {
-                if(err) return reject(err)
+                if (err) return reject(err)
                 resolve()
             })
         })
